master
I came across a cool new operator in java.util.Collectors. Its called teeing collector.
It was introduced in java 12 and facepalm me I just came to know about it now.

It can be used to collect the results of two collectors and return a result of a third collector.
Very useful when you want to collect the results of running two different operations on the same data
stream and return a result of a third collector.

I created a few basic examples of how to use it - trust me, it is so much fun to use.